requirements.txt
flask==2.0.1
flask-mysqldb==0.2.0
python-dotenv==0.19.0
werkzeug==2.0.1
flask-login==0.5.0